pyproject.toml
[project]
name = "study-mcp-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "MOON-HILL\\tsuki", email = "tsukioka.business@gmail.com" }
]
dependencies = [
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "openai>=2.6.1",
    "pygithub>=2.8.1",
    "mcp>=1.19.0",
    "gradio>=5.49.1",
    "anthropic>=0.71.0",
    "ruff>=0.14.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/study_mcp_server"]

[tool.poetry.dependencies]
pydantic = ">=2.0,<2.12"
gradio = "5.49.1"
